refactor(comment): fix some problems
