+++ src/comments/comment.controller.ts
@@ -8,17 +8,16 @@ import {
   Param,
   ParseIntPipe,
   Post,
-  Put,
   Query,
   UseFilters,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
+import { UpdateAttitudeRespondDto } from '../attitude/DTO/update-attitude.dto';
 import { parseAttitude } from '../attitude/attitude.enum';
 import { AuthService, AuthorizedAction } from '../auth/auth.service';
 import { BaseRespondDto } from '../common/DTO/base-respond.dto';
 import { BaseErrorExceptionFilter } from '../common/error/error-filter';
-import { AttitudeCommentDto } from './DTO/attitude-comment.dto';
 import { CreateCommentResponseDto } from './DTO/create-comment.dto';
 import { GetCommentDetailResponseDto } from './DTO/get-comment-detail.dto';
 import { GetCommentsResponseDto } from './DTO/get-comments.dto';
@@ -72,6 +71,36 @@ export class CommentsController {
     };
   }
 
+  // This method must be put above createComment()
+  // to avoid routing ambiguous problem
+  @Post('/:commentId/attitudes')
+  async updateAttitudeToComment(
+    @Param('commentId', ParseIntPipe) commentId: number,
+    @Body('attitude_type') attitude: string,
+    @Headers('Authorization') auth: string | undefined,
+  ): Promise<UpdateAttitudeRespondDto> {
+    const userId = this.authService.verify(auth).userId;
+    this.authService.audit(
+      auth,
+      AuthorizedAction.other,
+      await this.commentsService.getCommentCreatedById(commentId),
+      'comment/attitude',
+      commentId,
+    );
+    const attitudes = await this.commentsService.setAttitudeToComment(
+      commentId,
+      userId,
+      parseAttitude(attitude),
+    );
+    return {
+      code: 200,
+      message: 'You have expressed your attitude towards the comment',
+      data: {
+        attitudes,
+      },
+    };
+  }
+
   @Post('/:commentableType/:commentableId')
   async createComment(
     @Param('commentableType')
@@ -120,34 +149,6 @@ export class CommentsController {
     return {
       code: 204,
       message: 'Comment deleted already',
-    };
-  }
-
-  @Put('/:commentId/attitude')
-  async attitudeToComment(
-    @Param('commentId', ParseIntPipe) commentId: number,
-    @Body('attitude_type') attitude: string,
-    @Headers('Authorization') auth: string | undefined,
-  ): Promise<AttitudeCommentDto> {
-    const userId = this.authService.verify(auth).userId;
-    this.authService.audit(
-      auth,
-      AuthorizedAction.other,
-      await this.commentsService.getCommentCreatedById(commentId),
-      'comment/attitude',
-      commentId,
-    );
-    const attitudes = await this.commentsService.setAttitudeToComment(
-      commentId,
-      userId,
-      parseAttitude(attitude),
-    );
-    return {
-      code: 200,
-      message: 'You have expressed your attitude towards the comment',
-      data: {
-        attitudes,
-      },
     };
   }
